+++ docs/CHANGES.rst
@@ -11,6 +11,8 @@ e.g. performance).
 - ``Iterable.islice`` will raise a more appropriate ``TypeError`` when called
   without arguments.
 
+- ``partial`` only allows plain ``str`` as keyword-names in CPython 3.8.
+
 
 Version 0.7.0 (2018-01-28)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^